refactor(acp): extract transcript row model
